+++ mek_stripe_terminal/ios/mek_stripe_terminal/Sources/mek_stripe_terminal/Mappings/EasyConnectConfigurationMappings.swift
@@ -1,0 +1,12 @@
+
+import StripeTerminal
+
+extension EasyConnectConfigurationApi {
+    func toHost() throws -> EasyConnectConfiguration? {
+        switch self {
+
+        default:
+            fatalError()
+        }
+    }
+}